Attempt to deal with transparent unions
LuaJIT does not deal with the GNU transparent_union extension.
Unfortunately, that's what you get when you preprocess glibc header files
with _GNU_SOURCE.  Try to deal with these by replacing the inner union
with the first field.  However, sometimes the first field is not the
correct one (i.e. wait(2)), so also check a blacklist.